new formula: rubberband 1.5.0
Rubber Band Library is a high quality software library for audio
time-stretching and pitch-shifting. It permits you to change the tempo
and pitch of an audio stream or recording dynamically and independently
of one another.

includes ia command-line utility program that you can use for fixed
adjustments to the speed and pitch of existing audio files.